References issue #31
Make default LoggerFactory available on LFService as LFService.DEFAULT
This LoggerFactory can be used and shared among different projects
and allow an 'end application' to configure the logging for everything at once.
Library developers e.a. are encouraged to use this LoggerFactory if they do use the
log4j variant of logging in their library.

Added some tests to cover (and fixed some tests that were not compiling due to previous refactor).