Issue #43: Fix functions not defined on "window" object in ExtensionHelper
Make sure to check both window and the actual functions called upon
the window object. Some libraries polyfill the functions, others don't.
Additionally only set the registered flag, when the function are called
successfully (better for performance).

In case of react native there is no point to use these functions even,
since that is all for the chrome extension anyway.

Bump version to 0.6.4.