Refers issue #68
Split the config in a normal and production one for tsconfig.

This is necessary as jest (and possibly other dev dependencies) load in
a newer ES version on the fly, which makes the one set (ES2015) become newer and
worse allow newer things to be used. This opens serious problems as we
don't want to use newer things and ship polyfills to keep things working for
all browsers.

So the tests compile with the normal tsconfig, but the actual (production)
build uses the prod tsconfig to assure we get compile errors.