Refers issue #68
Add an identifier to the Log4TSGroupConfig, this is optional but when
used will allow easier control in the future (related to the log control).

Add updateRuntimeSettingsGroups and updateRuntimeSettings to Log4TSProvider
to allow updating the log level for groups and loggers as well as
changing the channel (by updateRuntimeSettings).