Issue #77: Change the internal logging exposure from core
Put it behind a separate namespace $internal, it should be very clear
then that it's not supposed to be used by end users.

Additionally add Log4TSProvider.clear() to make testing easier as it
allows a full/clear reset of the provider.

Contains a bit of reformatting as well (no changes there).