Issue #77: Remove the rollover from retention directly
Instead pass it in as additional NodeChannelOptions as it's really not
something most users need.

Additionally a RetentionStrategy is passed the rollover on

Implement rollover correctly, that we only call the function after a
file was flushed and closed (so not immediately when the maxfiles retention
is about to rollover as the file is not flushed/closed then).