For future reference, it's unclear on how exactly we want to split core and the actual implementation modules. Make typescript-logging into typescript-logging-core.
So we do all imports from typescript-logging-core (in case we do want
to publish it separately on npm later) for depending modules (log4ts, category).

The current setup is a mix between workspaces and ordinary npm link. Not
the prettiest, but good enough until some final decision
is made in the future.

Add/update various core config files.